zhihu/03/zhihu/requirements.txt
scrapy = 1.6.0
Pillow >= 5.0.0
matplotlib >= 2.1.2
requests >= 2.18.4
pyexecjs >= 1.5.1
